Hello World application from last night
This does not include headers & libraries for POSIX-UEFI, or the final PE build and GPT/FAT32 formatted image files for booting.

Docker environment exists to isolate utilities and cross-compiler toolchain, but the Docker file is incomplete (never finished it last night, just committed an image from a live container instead).